grpcclient/requirements.txt
fastapi>=0.111.0,<0.113.0
pydantic>=2.7.0,<3.0.0
fastapi[standard]
grpcio==1.67.0
grpcio-tools==1.67.0
azure-mgmt-applicationinsights
azure-identity

azure-monitor-opentelemetry==1.6.2
opentelemetry-instrumentation-fastapi==0.48b0